Remove JSVirtualMachine.
It doesn't make a lot of sense as it constraints to +1 JSContexts but then doesn't provide space to a real possibility of no-contexts, which means no machine (because a context group isn't a vm by its own).

As part of this changes I also try to provide a new memory model for JSContext where it can become clonable relying internally in JSC's retain/release mechanisms.