Make JSContext Send.
This is very likely to be true, although it also needs to come with some sort of sync. There's a small-ish detail that makes this yet wrong and that's the existance of JSContext::from for the inner reference. I need to figure out where to put the mutex reference in a way that can be recovered by the callback macro.