Adding an internal JSProtectedValue implementation.
It's internal only yet because I think this can have a much better interface, and probably be generic to the type it's protecting (JSValue / JSObject) so it's not lost. On the other hand, to guarantee the protection we need to clone JSContext and that's unsafe because of the unsafe impl Send. So for now it's internal so I can consume it in JSPromise. Might make unsafe public or at least importable through use foo_jsc::internal::JSProtectedValue or something like that?